third_party/blink/web_tests/virtual/saa-beyond-cookies-disabled/external/wpt/workers/same-site-cookies/third-party.all.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Check SharedWorker sameSiteCookies option all for third-party
  assert_equals: Worker should not have started expected "DidNotStart" but got "ReadOnLoad:None,ReadOnFetch:None"
Harness: the test ran to completion.
